Refactoring: Uso de Configuracion para no hardcodear parametros. Tambien uso la misma clase Semaphore del ej4, que es mas flexible

+++ trunk/Ejercicio 3/src/iniciador.cpp
@@ -7,32 +7,38 @@
  * Crea todos los IPCs a usar e inicia todos los procesos correspondientes a la aplicacion
  */
 
+#include "logger/Logger.h"
+#include "common/AtendedorDispositivos.h"
+#include "common/Planilla.h"
+#include "common/DespachadorTecnicos.h"
+#include "common/Configuracion.h"
+#include "ipc/Semaphore.h"
 #include <cstdlib>
 #include <fstream>
 #include <common/common.h>
 #include "unistd.h"
 #include <errno.h>
 
-#include "ipc/Semaphore.h"
-#include "logger/Logger.h"
-#include "common/AtendedorDispositivos.h"
-#include "common/Planilla.h"
-#include "common/DespachadorTecnicos.h"
-
 /*
  * 
  */
 
-void createIPCObjects();
-void createSystemProcesses();
+void createIPCObjects( const Configuracion& config );
+void createSystemProcesses( const Configuracion& config );
 
 int main(int argc, char** argv) {
 
-    Logger::initialize(logFileName.c_str(), Logger::LOG_DEBUG);
+    Logger::initialize(Constantes::ARCHIVO_LOG.c_str(), Logger::LOG_DEBUG);
     Logger::error("Logger inicializado. Inicializando IPCs...", __FILE__);
 
+    Configuracion config;
+    if( !config.LeerDeArchivo() ){
+        Logger::error("Archivo de configuracion no encontrado", __FILE__);
+        return 1;
+    }
+
     try {
-        createIPCObjects();
+        createIPCObjects( config );
     } catch(std::string err) {
         Logger::error("Error al crear los objetos activos...", __FILE__);
         Logger::destroy();
@@ -40,7 +46,7 @@ int main(int argc, char** argv) {
     }
     Logger::debug("Objetos IPC inicializados correctamente. Iniciando procesos...", __FILE__);
 
-    createSystemProcesses();
+    createSystemProcesses( config );
     Logger::debug("Procesos iniciados correctamente...", __FILE__);
     
     Logger::notice("Sistema inicializado correctamente...", __FILE__);
@@ -50,10 +56,11 @@ int main(int argc, char** argv) {
     return 0;
 }
 
-void createIPCObjects() {
+void createIPCObjects( const Configuracion& config ) {
+    const std::string archivoIpcs = config.ObtenerParametroString( Constantes::NombresDeParametros::ARCHIVO_IPCS.c_str() );
 
     // Creo el archivo que se usara para obtener las keys
-    std::fstream ipcFile(ipcFileName.c_str(), std::ios::out);
+    std::fstream ipcFile(Constantes::ARCHIVO_LOG.c_str(), std::ios::out);
     if (ipcFile.bad() || ipcFile.fail()) {
         std::string err = std::string("Error creando el archivo de IPCs. Error: ") + std::string(strerror(errno));
         Logger::error(err.c_str(), __FILE__);
@@ -62,15 +69,17 @@ void createIPCObjects() {
     ipcFile.close();
 
     // Colas de mensajes entre dispositivo y testers
-    AtendedorDispositivos atendedor;
+    AtendedorDispositivos atendedor( config );
 
     // Creo semaforo para la shmem general de la planilla
-    Semaphore semPlanilla(SEM_PLANILLA_GENERAL);
+    Semaphore semPlanilla( archivoIpcs, config.ObtenerParametroEntero( Constantes::NombresDeParametros::SEM_PLANILLA_GENERAL )  );
     semPlanilla.creaSem();
     semPlanilla.iniSem(1); // Inicializa el semaforo en 1
 
     // Creo las shmem y semaforos de las planillas
-    for( int iTester=ID_TESTER_START; iTester<ID_TESTER_START+CANT_TESTERS;iTester++ ){
+    int iTesterStart = config.ObtenerParametroEntero( Constantes::NombresDeParametros::ID_TESTER_START );
+    int cantTesters = config.ObtenerParametroEntero( Constantes::NombresDeParametros::CANT_TESTERS );
+    for( int iTester=iTesterStart; iTester < iTesterStart + cantTesters; iTester++ ){
         Planilla planilla( iTester );
     }
 
@@ -79,25 +88,12 @@ void createIPCObjects() {
 
 }
 
-void createSystemProcesses() {
-
-    // Creo dispositivos
-    for(int i = 0; i < CANT_DISPOSITIVOS; i++) {
-        char param[3];
-        int idDispositivo = ID_DISPOSITIVO_START + i;
-        sprintf(param, "%d\n", idDispositivo);
-        pid_t newPid = fork();
-        if(newPid == 0) {
-            // Inicio el programa correspondiente
-            execlp("./dispositivo", "dispositivo", param, (char*)0);
-            Logger::error("Error al ejecutar el programa dispositivo de ID" + idDispositivo, __FILE__);
-        }
-    }
+void createSystemProcesses( const Configuracion& config ) {
 
     // Creo testers
-    for(int i = 0; i < CANT_TESTERS; i++) {
+    for(int i = 0; i < config.ObtenerParametroEntero( Constantes::NombresDeParametros::CANT_TESTERS ); i++) {
         char param[3];
-        int idTester = ID_TESTER_START + i;
+        int idTester = config.ObtenerParametroEntero( Constantes::NombresDeParametros::ID_TESTER_START ) + i;
         sprintf(param, "%d\n", idTester);
         pid_t newPid = fork();
         if(newPid == 0) {
@@ -112,6 +108,19 @@ void createSystemProcesses() {
     if(tecPid == 0) {
         execlp("./tecnico", "tecnico", (char*)0);
         Logger::error("Error al ejecutar el programa tecnico", __FILE__);
+    }
+    
+    // Creo dispositivos
+    for(int i = 0; i < config.ObtenerParametroEntero( Constantes::NombresDeParametros::CANT_DISPOSITIVOS ); i++) {
+        char param[3];
+        int idDispositivo = config.ObtenerParametroEntero( Constantes::NombresDeParametros::ID_DISPOSITIVO_START ) + i;
+        sprintf(param, "%d\n", idDispositivo);
+        pid_t newPid = fork();
+        if(newPid == 0) {
+            // Inicio el programa correspondiente
+            execlp("./dispositivo", "dispositivo", param, (char*)0);
+            Logger::error("Error al ejecutar el programa dispositivo de ID" + idDispositivo, __FILE__);
+        }
     }
 
     Logger::debug("Programas iniciados correctamente...", __FILE__);
